# Make THStorage / THCStorage have void* data ptr.