--- a/aten/src/ATen/templates/StorageDerived.cpp
+++ b/aten/src/ATen/templates/StorageDerived.cpp
@@ -113,11 +113,11 @@ std::size_t ${Storage}::size() const {
 }
 
 void* ${Storage}::data() {
-  return storage->data;
+  return storage->data_ptr;
 }
 
 const void* ${Storage}::data() const {
-  return storage->data;
+  return storage->data_ptr;
 }
 
 auto ${Storage}::retain() -> ${Storage}& {
@@ -164,7 +164,7 @@ auto ${Storage}::get(std::size_t ind) -> Scalar {
 auto ${Storage}::fast_get(std::size_t ind) -> Scalar {
   if(${isCUDA})
     throw std::runtime_error("unsupported operation 'fast_get'");
-  return static_cast<${ScalarType}>(${to_at_type}(storage->data[ind]));
+  return static_cast<${ScalarType}>(${to_at_type}(storage->unsafeData<${THScalarType}>()[ind]));
 }
 
 void ${Storage}::set_flag(char flag) {