--- a/aten/src/TH/THStorage.hpp
+++ b/aten/src/TH/THStorage.hpp
@@ -5,6 +5,9 @@
 
 #include "THStorage.h"
 
+#include "ATen/ScalarType.h"
+#include "ATen/ScalarTypeUtils.h"
+#include "THTypeConversion.hpp"
 #include <atomic>
 
 #include "generic/THStorage.hpp"