--- a/aten/src/TH/generic/THStorage.cpp
+++ b/aten/src/TH/generic/THStorage.cpp
@@ -6,7 +6,7 @@
 
 real* THStorage_(data)(const THStorage *self)
 {
-  return self->data;
+  return self->data<real>();
 }
 
 ptrdiff_t THStorage_(size)(const THStorage *self)
@@ -34,7 +34,8 @@ THStorage* THStorage_(newWithAllocator)(ptrdiff_t size,
                                         void *allocatorContext)
 {
   THStorage *storage = static_cast<THStorage*>(THAlloc(sizeof(THStorage)));
-  storage->data = static_cast<real*>(allocator->malloc(allocatorContext, sizeof(real)*size));
+  storage->scalar_type = at::CTypeToScalarType<th::from_type<real>>::to();
+  storage->data_ptr = allocator->malloc(allocatorContext, sizeof(real)*size);
   storage->size = size;
   new (&storage->refcount) std::atomic<int>(1);
   storage->flag = TH_STORAGE_REFCOUNTED | TH_STORAGE_RESIZABLE | TH_STORAGE_FREEMEM;
@@ -62,34 +63,38 @@ THStorage* THStorage_(newWithMapping)(const char *filename, ptrdiff_t size, int 
 THStorage* THStorage_(newWithSize1)(real data0)
 {
   THStorage *self = THStorage_(newWithSize)(1);
-  self->data[0] = data0;
+  real *data = THStorage_(data)(self);
+  data[0] = data0;
   return self;
 }
 
 THStorage* THStorage_(newWithSize2)(real data0, real data1)
 {
   THStorage *self = THStorage_(newWithSize)(2);
-  self->data[0] = data0;
-  self->data[1] = data1;
+  real *data = THStorage_(data)(self);
+  data[0] = data0;
+  data[1] = data1;
   return self;
 }
 
 THStorage* THStorage_(newWithSize3)(real data0, real data1, real data2)
 {
   THStorage *self = THStorage_(newWithSize)(3);
-  self->data[0] = data0;
-  self->data[1] = data1;
-  self->data[2] = data2;
+  real *data = THStorage_(data)(self);
+  data[0] = data0;
+  data[1] = data1;
+  data[2] = data2;
   return self;
 }
 
 THStorage* THStorage_(newWithSize4)(real data0, real data1, real data2, real data3)
 {
   THStorage *self = THStorage_(newWithSize)(4);
-  self->data[0] = data0;
-  self->data[1] = data1;
-  self->data[2] = data2;
-  self->data[3] = data3;
+  real *data = THStorage_(data)(self);
+  data[0] = data0;
+  data[1] = data1;
+  data[2] = data2;
+  data[3] = data3;
   return self;
 }
 
@@ -132,7 +137,7 @@ void THStorage_(free)(THStorage *storage)
     if(--storage->refcount == 0)
     {
       if(storage->flag & TH_STORAGE_FREEMEM) {
-        storage->allocator->free(storage->allocatorContext, storage->data);
+        storage->allocator->free(storage->allocatorContext, THStorage_(data)(storage));
       }
       if(storage->flag & TH_STORAGE_VIEW) {
         THStorage_(free)(storage->view);
@@ -153,7 +158,8 @@ THStorage* THStorage_(newWithDataAndAllocator)(real* data, ptrdiff_t size,
                                                THAllocator* allocator,
                                                void* allocatorContext) {
   THStorage *storage = static_cast<THStorage*>(THAlloc(sizeof(THStorage)));
-  storage->data = data;
+  storage->scalar_type = at::CTypeToScalarType<th::from_type<real>>::to();
+  storage->data_ptr = data;
   storage->size = size;
   storage->refcount = 1;
   storage->flag = TH_STORAGE_REFCOUNTED | TH_STORAGE_RESIZABLE | TH_STORAGE_FREEMEM;
@@ -168,14 +174,14 @@ void THStorage_(resize)(THStorage *storage, ptrdiff_t size)
   {
     if(storage->allocator->realloc == NULL) {
       /* case when the allocator does not have a realloc defined */
-      real *old_data = storage->data;
+      real *old_data = THStorage_(data)(storage);
       ptrdiff_t old_size = storage->size;
       if (size == 0) {
-        storage->data = NULL;
+        storage->data_ptr = NULL;
       } else {
-        storage->data = static_cast<real*>(storage->allocator->malloc(
+        storage->data_ptr = storage->allocator->malloc(
             storage->allocatorContext,
-            sizeof(real)*size));
+            sizeof(real)*size);
       }
       storage->size = size;
       if (old_data != NULL) {
@@ -184,15 +190,15 @@ void THStorage_(resize)(THStorage *storage, ptrdiff_t size)
           copy_size = storage->size;
         }
         if (copy_size > 0) {
-          memcpy(storage->data, old_data, sizeof(real)*copy_size);
+          memcpy(THStorage_(data)(storage), old_data, sizeof(real)*copy_size);
         }
         storage->allocator->free(storage->allocatorContext, old_data);
       }
     } else {
-      storage->data = static_cast<real*>(storage->allocator->realloc(
+      storage->data_ptr = storage->allocator->realloc(
               storage->allocatorContext,
-              storage->data,
-              sizeof(real)*size));
+              THStorage_(data)(storage),
+              sizeof(real)*size);
       storage->size = size;
     }
   } else {
@@ -204,32 +210,32 @@ void THStorage_(fill)(THStorage *storage, real value)
 {
   ptrdiff_t i;
   for(i = 0; i < storage->size; i++)
-    storage->data[i] = value;
+    THStorage_(data)(storage)[i] = value;
 }
 
 void THStorage_(set)(THStorage *self, ptrdiff_t idx, real value)
 {
   THArgCheck((idx >= 0) && (idx < self->size), 2, "out of bounds");
-  self->data[idx] = value;
+  THStorage_(data)(self)[idx] = value;
 }
 
 real THStorage_(get)(const THStorage *self, ptrdiff_t idx)
 {
   THArgCheck((idx >= 0) && (idx < self->size), 2, "out of bounds");
-  return self->data[idx];
+  return THStorage_(data)(self)[idx];
 }
 
 void THStorage_(swap)(THStorage *storage1, THStorage *storage2)
 {
 #define SWAP(val) { val = storage1->val; storage1->val = storage2->val; storage2->val = val; }
-    real *data;
+    void *data_ptr;
     ptrdiff_t size;
     char flag;
     THAllocator *allocator;
     void *allocatorContext;
     struct THStorage *view;
 
-    SWAP(data);
+    SWAP(data_ptr);
     SWAP(size);
     SWAP(flag);
     // don't swap refcount!