--- a/aten/src/TH/generic/THStorage.hpp
+++ b/aten/src/TH/generic/THStorage.hpp
@@ -12,19 +12,19 @@ typedef struct THStorage
     THAllocator *allocator;
     void *allocatorContext;
     struct THStorage *view;
-    template <typename T>
 
+    template <typename T>
     inline T * data() const {
       auto scalar_type_T = at::CTypeToScalarType<th::from_type<T>>::to();
       if (scalar_type != scalar_type_T) {
         AT_ERROR("Attempt to access Storage having data type ", at::toString(scalar_type),
                  " as data type ", at::toString(scalar_type_T));
       }
-      return unsafeData<T>();
+      return unsafe_data<T>();
     }
 
     template <typename T>
-    inline T * unsafeData() const {
+    inline T * unsafe_data() const {
       return static_cast<T*>(this->data_ptr);
     }
 } THStorage;