--- a/aten/src/TH/generic/THStorageCopy.cpp
+++ b/aten/src/TH/generic/THStorageCopy.cpp
@@ -5,22 +5,28 @@
 void THStorage_(rawCopy)(THStorage *storage, real *src)
 {
   ptrdiff_t i;
+  real *data = THStorage_(data)(storage);
   for(i = 0; i < storage->size; i++)
-    storage->data[i] = src[i];
+    data[i] = src[i];
 }
 
 void THStorage_(copy)(THStorage *storage, THStorage *src)
 {
   THArgCheck(storage->size == src->size, 2, "size mismatch");
-  THStorage_(rawCopy)(storage, src->data);
+  THStorage_(rawCopy)(storage, THStorage_(data)(src));
 }
+
+// NOTE: for performance, these macros generally directly access data, rather
+// than THStorage_(data)ß
 
 #define IMPLEMENT_THStorage_COPY(TYPENAMESRC) \
 void THStorage_(copy##TYPENAMESRC)(THStorage *storage, TH##TYPENAMESRC##Storage *src) \
 { \
-  ptrdiff_t i;                                                        \
-  for(i = 0; i < storage->size; i++)                                  \
-    storage->data[i] = (real)src->data[i];                            \
+  ptrdiff_t i;                                                          \
+  auto data = THStorage_(data)(storage);                                \
+  auto src_data = TH##TYPENAMESRC##Storage_data(src);                   \
+  for(i = 0; i < storage->size; i++)                                    \
+    data[i] = static_cast<real>(src_data[i]);                           \
 }
 
 #define IMPLEMENT_THStorage_COPY_FROM_HALF(TYPENAMESRC)		\
@@ -28,8 +34,10 @@ void THStorage_(copy##TYPENAMESRC)(THStorage *storage, TH##TYPENAMESRC##Storage 
 { \
   THArgCheck(storage->size == src->size, 2, "size mismatch"); \
   ptrdiff_t i;								\
+  auto data = THStorage_(data)(storage);      \
+  auto src_data = TH##TYPENAMESRC##Storage_data(src); \
   for(i = 0; i < storage->size; i++)					\
-    storage->data[i] = (real)TH_half2float(src->data[i]);		\
+    data[i] = (real)TH_half2float(src_data[i]); \
 }
 
 #define IMPLEMENT_THStorage_COPY_TO_HALF(TYPENAMESRC)		\
@@ -37,8 +45,10 @@ void THStorage_(copy##TYPENAMESRC)(THStorage *storage, TH##TYPENAMESRC##Storage 
 { \
   THArgCheck(storage->size == src->size, 2, "size mismatch"); \
   ptrdiff_t i;								\
+  auto data = THStorage_(data)(storage);      \
+  auto src_data = TH##TYPENAMESRC##Storage_data(src); \
   for(i = 0; i < storage->size; i++)					\
-    storage->data[i] = TH_float2half((float)(src->data[i]));		\
+    data[i] = TH_float2half((float)(src_data[i])); \
 }
 
 #define IMPLEMENT_THStorage_COPY_TO_FROM_HALF(TYPENAMESRC)		\
@@ -46,8 +56,10 @@ void THStorage_(copy##TYPENAMESRC)(THStorage *storage, TH##TYPENAMESRC##Storage 
 { \
   THArgCheck(storage->size == src->size, 2, "size mismatch"); \
   ptrdiff_t i;								\
+  auto data = THStorage_(data)(storage);      \
+  auto src_data = TH##TYPENAMESRC##Storage_data(src); \
   for(i = 0; i < storage->size; i++)					\
-    storage->data[i] = src->data[i];		\
+    data[i] = static_cast<real>(src_data[i]); \
 }
 
 #ifndef TH_REAL_IS_HALF