@gchanan 2018-05-31
A couple of microbenchmarks (accessing storage and copying storage):

before this patch:
```
In [5]: %t x[0]
47.2 ns ± 1.44 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [13]: %t x.copy_(y)
338 ns ± 8.41 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```
after this patch:
```
In [5]: %t x[0]
46.3 ns ± 1.55 ns per loop (mean ± std. dev. of 7 runs, 10000000 loops each)

In [6]:  %t x.copy_(y)
326 ns ± 8.66 ns per loop (mean ± std. dev. of 7 runs, 1000000 loops each)
```

so within the margin of error.